+++ internal/chunk/readjob.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/getsentry/vroom/internal/nodetree"
 	"github.com/getsentry/vroom/internal/storageutil"
+	"github.com/getsentry/vroom/internal/utils"
 	"gocloud.dev/blob"
 )
 
@@ -20,6 +21,7 @@ type (
 		ThreadID       *string
 		Start          uint64
 		End            uint64
+		Intervals      map[string][]utils.Interval
 		Result         chan<- storageutil.ReadJobResult
 	}
 
@@ -30,6 +32,7 @@ type (
 		ThreadID      *string
 		Start         uint64
 		End           uint64
+		Intervals     map[string][]utils.Interval
 	}
 )
 
@@ -50,6 +53,7 @@ func (job ReadJob) Read() {
 		ThreadID:      job.ThreadID,
 		Start:         job.Start,
 		End:           job.End,
+		Intervals:     job.Intervals,
 	}
 }
 
@@ -68,6 +72,7 @@ type (
 		ThreadID      *string
 		Start         uint64
 		End           uint64
+		Intervals     map[string][]utils.Interval
 	}
 )
 
@@ -86,8 +91,31 @@ func (job CallTreesReadJob) Read() {
 		return
 	}
 
-	callTrees, err := chunk.CallTrees(job.ThreadID)
-
+	callTrees := make(map[string][]*nodetree.Node)
+	for tid := range job.Intervals {
+		if tid == "" {
+			callTrees, err = chunk.CallTrees(nil)
+			if err != nil {
+				job.Result <- CallTreesReadJobResult{Err: err}
+				return
+			}
+			// we've already computed the callTrees for all the
+			// tids that we might need so we can bail out
+			break
+		}
+		if _, ok := callTrees[tid]; ok {
+			// if a former interval already had the same
+			// tid we've already computed that callTree
+			// and we can bail out early
+			continue
+		}
+		callTree, err := chunk.CallTrees(&tid)
+		if err != nil {
+			job.Result <- CallTreesReadJobResult{Err: err}
+			return
+		}
+		callTrees[tid] = callTree[tid]
+	}
 	job.Result <- CallTreesReadJobResult{
 		Err:           err,
 		CallTrees:     callTrees,