+++ internal/flamegraph/span_util.go
@@ -2,35 +2,11 @@ package flamegraph
 
 import (
 	"math"
-	"sort"
 	"time"
 
 	"github.com/getsentry/vroom/internal/nodetree"
 	"github.com/getsentry/vroom/internal/utils"
 )
-
-func mergeIntervals(intervals *[]utils.Interval) []utils.Interval {
-	if len(*intervals) == 0 {
-		return *intervals
-	}
-	sort.SliceStable((*intervals), func(i, j int) bool {
-		if (*intervals)[i].Start == (*intervals)[j].Start {
-			return (*intervals)[i].End < (*intervals)[j].End
-		}
-		return (*intervals)[i].Start < (*intervals)[j].Start
-	})
-
-	newIntervals := []utils.Interval{(*intervals)[0]}
-	for _, interval := range (*intervals)[1:] {
-		if interval.Start <= newIntervals[len(newIntervals)-1].End {
-			newIntervals[len(newIntervals)-1].End = max(newIntervals[len(newIntervals)-1].End, interval.End)
-		} else {
-			newIntervals = append(newIntervals, interval)
-		}
-	}
-
-	return newIntervals
-}
 
 func relativeIntervalsFromAbsoluteTimestamp(intervals *[]utils.Interval, t uint64) {
 	for i, v := range *intervals {
@@ -44,20 +20,21 @@ func relativeIntervalsFromAbsoluteTimestamp(intervals *[]utils.Interval, t uint6
 func sliceCallTree(callTree *[]*nodetree.Node, intervals *[]utils.Interval) []*nodetree.Node {
 	slicedTree := make([]*nodetree.Node, 0)
 	for _, node := range *callTree {
+		newNode := *node
 		if duration := getTotalOverlappingDuration(node, intervals); duration > 0 {
 			sampleCount := int(math.Ceil(float64(duration) / float64(time.Millisecond*10)))
 			// here we take the minimum between the node sample count and the estimated
 			// sample count to mitigate the case when we make a wrong estimation due
 			// to sampling frequency not being respected. (Python native code holding
 			// the GIL, php, etc.)
-			node.SampleCount = min(sampleCount, node.SampleCount)
-			node.DurationNS = duration
+			newNode.SampleCount = min(sampleCount, node.SampleCount)
+			newNode.DurationNS = duration
 			if children := sliceCallTree(&node.Children, intervals); len(children) > 0 {
-				node.Children = children
+				newNode.Children = children
 			} else {
-				node.Children = nil
+				newNode.Children = nil
 			}
-			slicedTree = append(slicedTree, node)
+			slicedTree = append(slicedTree, &newNode)
 		}
 	} // end range callTree
 	return slicedTree